feat(infra): introduce docker containerization and multi-container orchestration
- feat(infra): add multi-stage Dockerfile with restricted non-root user and Gunicorn preloading

- feat(infra): add docker-compose.yml to orchestrate web runtime with a healthy Postgres backend

- fix(api): cast numpy.float64 model outputs to native floats to resolve psycopg2 database insertion failures

- fix(ui): replace empty template placeholder with actual LinkedIn profile URL in footer

- docs: rewrite README.md to prioritize container deployment and document the REST API schema

- chore: update .gitignore to isolate deployment readmes and refresh LICENSE copyright details

+++ app/api.py
@@ -43,7 +43,7 @@ def predict_delivery():
 
         # 3. Now I can get the prediction.
         prediction = model.predict(input_data)[0]
-        estimated_hours = round(prediction, 2)
+        estimated_hours = float(round(prediction, 2))
 
         # 4. This is new: I'm saving the prediction to my database.
         new_log = PredictionLog(
